Limit GetData()  refresh rate and cache  some data

monsters : 200ms 
Inventory : 500ms
Objects :  200ms

Everything else (important core data) is instant refresh so its following the 100ms ticker on background loop or instant if requested from other part of code.

Note that hovered data for inventory (item) still get updated to prevent issue with   different transactions. ( pickup, vendor,stash etc)